requirements.txt
requests
asyncio
bs4
discord
datetime
firebase_admin
quickchart.io
nest_asyncio
table2ascii
google-generativeai